+++ DbLibrary/StatisticModel/PersonalDeathByStat.cs
@@ -1,26 +1,24 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLibrary.DbContexts
+namespace DbLibrary.StatisticModel
 {
-    [Table("personaldeathbyweaponstat")]
-    public class PersonalDeathByWeaponStat
+    [Table("personaldeathbystat")]
+    public class PersonalDeathByStat
     {
         [Key]
-        [Column("PdbwsID")]
+        [Column("PdbsID")]
         public ulong ID { get; set; }
 
         [Column("PmsID")]
         public ulong PersonalMatchStatID { get; set; }
         public virtual PersonalMatchStat? PersonalMatchStat { get; set; }
 
-        [Column("WeaponID")]
-        public ushort WeaponID { get; set; }
-        public virtual Weapon? Weapon { get; set; }
+        [Column("SteamProfileID")]
+        public uint SteamProfileID { get; set; }
+        public virtual SteamProfile? SteamProfile { get; set; }
 
         [Column("Count")]
         public ushort Count { get; set; }
-
-
     }
 }
