+++ DbLibrary/StatisticModel/PersonalDeathByWeaponStat.cs
@@ -1,28 +1,25 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLibrary.DbContexts
+namespace DbLibrary.StatisticModel
 {
-    [Table("personalkillstat")]
-    public class PersonalKillStat
+    [Table("personaldeathbyweaponstat")]
+    public class PersonalDeathByWeaponStat
     {
         [Key]
-        [Column("PksID")]
+        [Column("PdbwsID")]
         public ulong ID { get; set; }
 
         [Column("PmsID")]
         public ulong PersonalMatchStatID { get; set; }
         public virtual PersonalMatchStat? PersonalMatchStat { get; set; }
 
-        [Column("SteamProfileID")]
-        public uint SteamProfileID { get; set; }
-
-        public virtual SteamProfile? SteamProfile { get; set; }
+        [Column("WeaponID")]
+        public ushort WeaponID { get; set; }
+        public virtual Weapon? Weapon { get; set; }
 
         [Column("Count")]
         public ushort Count { get; set; }
-
-
 
 
     }
