+++ DbLibrary/StatisticModel/PersonalKillStat.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLibrary.DbContexts
+namespace DbLibrary.StatisticModel
 {
-    [Table("personaldeathbystat")]
-    public class PersonalDeathByStat
+    [Table("personalkillstat")]
+    public class PersonalKillStat
     {
         [Key]
-        [Column("PdbsID")]
+        [Column("PksID")]
         public ulong ID { get; set; }
 
         [Column("PmsID")]
@@ -16,9 +16,14 @@ namespace DbLibrary.DbContexts
 
         [Column("SteamProfileID")]
         public uint SteamProfileID { get; set; }
+
         public virtual SteamProfile? SteamProfile { get; set; }
 
         [Column("Count")]
         public ushort Count { get; set; }
+
+
+
+
     }
 }
