+++ DbLibrary/StatisticModel/PersonalWeaponKillStat.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLibrary.DbContexts
+namespace DbLibrary.StatisticModel
 {
     [Table("personalweaponkillstat")]
     public class PersonalWeaponKillStat
