+++ DbLibrary/StatisticModel/SteamProfile.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLibrary.DbContexts
+namespace DbLibrary.StatisticModel
 {
     [Table("steamprofile")]
     public class SteamProfile
@@ -18,6 +18,9 @@ namespace DbLibrary.DbContexts
 
         [Column("AvatarHash")]
         public string? AvatarHash { get; set; }
+
+        [Column("ProfileUpdateTime")]
+        public DateTime? ProfileUpdtaded { get; set; }
 
         public virtual ICollection<PersonalKillStat>? KillStats { get; set; }
         public virtual ICollection<PersonalDeathByStat>? DeathByStats { get; set; }
