+++ WebApp/Services/RconScanerService/RconUpdaterService.cs
@@ -1,4 +1,4 @@
-﻿using DbLibrary.DbContexts;
+﻿using DbLibrary.StatisticModel;
 using Microsoft.EntityFrameworkCore;
 using System.Text.Json;
 using System.Text.RegularExpressions;
