+++ WebApp/WebApp.cs
@@ -1,4 +1,4 @@
-using DbLibrary.DbContexts;
+using DbLibrary.StatisticModel;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.EntityFrameworkCore;
 using System.Globalization;
